+++ CHANGELOG.md
@@ -8,10 +8,11 @@ All Notable changes to `Csv` will be documented in this file
 
 - `League\Csv\EncloseField` to force enclosure insertion on every field
 - `League\Csv\RFC4180Field::addTo` accept an option `$replace_whitespace` argument to improve RFC4180 compliance.
+- `League\Csv\Abstract::getContent`
 
 ### Deprecated
 
-- Nothing
+- `League\Csv\Abstract::__toString` use `League\Csv\Abstract::getContent` instead
 
 ### Fixed